requirements.txt
# Deepfake project dependencies
torch==2.9.0
torchvision==0.24.0
numpy>=2.3.0
pillow>=11.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
ImageHash>=4.3.1
scikit-learn>=1.7.2
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
scipy>=1.11.0
matplotlib>=3.9.0
joblib>=1.4.0
